Adding HipChat notifications for MDID workers
These are really quick-and-dirty scripts to try and stick a band-aid on
the MDID workers service.

The workers service crashes and needs to be restarted quite often.  I
found in the service properties that you can configure actions for the
first, second, and subsequent stopages.  So on the first/second times
it stops we'll restart the service and send a notification to HipChat.

On any subsequent stopages, we will just notify HipChat for now, so we
don't end up in some kind of loop where the service just keeps stopping
and restarting itself.