wfastcgi.py patched for faster file uploads
The wfastcgi.py interface wrote all of the STDIN input from FastCGI into
the `wsgi.input` WSGI parameter in-memory.  For large file uploads, this
made things go real slow (and also ate up memory).

Changed the `wsgi.input` WSGI parameter to be a SpooledTemporaryFile
rather than an in-memory object, and this increased performance
drastically.  The SpooledTemporaryFile will buffer data in-memory until
it reaches a max limit (set to 500k in our implementation) and once the
size goes over this limit it is spooled to a temporary file on disk.