Fix XPath query for HipChat event notifications
So the batch file that processes the events to send to the HipChat
notifier uses the wevtutil.exe util to pull event details.  The only
issue was that the XPath query was hard-coded to look in the 'System'
path, which works for the Service Control events, but the actual 'MDID
Workers' events come through the Application path.

The event_viewer_notifier.bat file receives the Channel as an argument,
so I'm using that to specify the Path attribute to the <Query> and
<Select> tags.