Handle dropped database connections for workers
Added some exception handling in the workers_callback() function
to catch an OperationalError, thrown when the MySQL server drops a
database connection.  The exception handler will use the
django.db.connection.close() method to close the db connection, forcing
Django to re-establish the connection the next time it needs one.  Then
the job message is requeued in the RabbitMQ stack, and will be picked
back up on the next iteration.