Got a working Vagrant VM for MDID development
The Vagrant VM is just about there.  It sets up the environment, port
forwarding is working, and settings are configured.  All that is needed
is to ssh into the VM (`vagrant ssh`) and then run the `./runserver`
script in the vagrant home directory to bring up the django development
server.  From there you can navigate to 'http://localhost:8888' on your
host machine and you should see MDID.

We can only get so far with generic customization of the
`settings_local.py` file, so manual customization is still needed to put
in specific credentials, e.g. for LDAP authentication.

WIP: Still to do is to write up an init script to start up Solr when the
box comes up, and put it in place and set the appropriate runlevels.