Fix the upstart scripts for mdid3 services
The mdid3-solr and mdid3-workers services were apparently not starting
on boot.  Turns out the issue was that these services run out of the
shared directory (/vagrant) mounted by Vagrant, but that directory isn't
mounted at the time the services were trying to start up.

But what could have been a big mess, turned out to be trivial...because
Vagrant and Upstart have our back on this one.  Vagrant emits an Upstart
event 'vagrant-mounted' when the shared directories have been mounted.
So I just needed to change the Upstart scripts for these services to
start on that event, rather than on a particular runlevel, and boom.