Add HTTP auth to the shared collections worker
The shared collections worker doesn't do any HTTP authentication, even
if a username/password is provided in the shared collection record.
This means that if the installation hosting the shared collection
requires authentication, then a 404 NOT FOUND error is returned when the
worker tries to retrieve the image.

This fix adds a check to see if there is a username and password set in
the shared collection record, and if so then it does an authenticated
request.  Otherwise it falls back to a non-authenticated request.

Also, I used the `requests` library rather than the `urllib2` lib.
Requests is just so much more painless to work with :)